# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, data, file, plugin, mcp, skill, plus the sensitive
    config:edit_admin and chat:admin sub-scopes. The config scope also grants bot
    and provider access, but no parent scope grants a sensitive sub-scope.
...
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/chat:
    post:
      tags: [Chat]
      summary: Send a webchat message
      operationId: sendChatMessage
      x-astrbot-scope: chat
      x-astrbot-sensitive-scopes: ["chat:admin"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /api/v1/chat/sessions:
    get:
      tags: [Chat]
      summary: List webchat sessions
      operationId: listChatSessions
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: username
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/new:
    get:
      tags: [Chat]
      summary: Create a webchat session
      operationId: createChatSession
      x-astrbot-scope: chat
      parameters:
        - name: platform_id
          in: query
          schema:
            type: string
            default: webchat
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/batch-delete:
    post:
      tags: [Chat]
      summary: Delete multiple webchat sessions
      operationId: batchDeleteChatSessions
      x-astrbot-scope: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionBatchDeleteRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/batch-archive:
    post:
      tags: [Chat]
      summary: Archive multiple webchat sessions
      operationId: batchArchiveChatSessions
      x-astrbot-scope: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionBatchDeleteRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/batch-unarchive:
    post:
      tags: [Chat]
      summary: Restore multiple archived webchat sessions
      operationId: batchUnarchiveChatSessions
      x-astrbot-scope: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionBatchDeleteRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/archived:
    get:
      tags: [Chat]
      summary: List archived webchat sessions
      operationId: listArchivedChatSessions
      x-astrbot-scope: chat
      parameters:
        - name: platform_id
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
            default: ""
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/{session_id}/archive:
    post:
      tags: [Chat]
      summary: Archive a webchat session
      operationId: archiveChatSession
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/{session_id}/unarchive:
    post:
      tags: [Chat]
      summary: Restore an archived webchat session
      operationId: unarchiveChatSession
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/{session_id}:
    get:
      tags: [Chat]
      summary: Get a webchat session and its messages
      operationId: getChatSession
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    patch:
      tags: [Chat]
      summary: Update a webchat session
      operationId: updateChatSession
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionPatchRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Chat]
      summary: Delete a webchat session
      operationId: deleteChatSession
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/{session_id}/stop:
    post:
      tags: [Chat]
      summary: Stop active webchat work for a session
      operationId: stopChatSession
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/open-file:
    post:
      tags: [Chat]
      summary: Open a local file on the AstrBot host with the OS default application
      operationId: openChatLocalFile
      x-astrbot-scope: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatOpenFileRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/runs/{run_id}/stream:
    get:
      tags: [Chat]
      summary: Resume an active webchat run as an SSE stream
      operationId: resumeChatRun
      x-astrbot-scope: chat
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resumed server-sent chat stream or an error envelope
  /api/v1/chat/sessions/{session_id}/messages/{message_id}:
    patch:
      tags: [Chat]
      summary: Update the latest user message in a webchat session
      operationId: updateChatMessage
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: message_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessagePatchRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/{session_id}/messages/{message_id}/regenerate:
    post:
      tags: [Chat]
      summary: Regenerate a bot message in a webchat session
      operationId: regenerateChatMessage
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: message_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageRegenerateRequest"
      responses:
        "200":
          description: Server-sent chat stream or an error envelope
  /api/v1/chat/sessions/{session_id}/messages/{message_id}/branch:
    post:
      tags: [Chat]
      summary: Branch a webchat session from a bot message
      operationId: branchChatMessage
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: message_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /api/v1/chat/threads:
    post:
      tags: [Chat]
      summary: Create or reuse a side thread from a webchat message
      operationId: createChatThread
      x-astrbot-scope: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatThreadCreateRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/threads/{thread_id}:
    get:
      tags: [Chat]
      summary: Get a webchat side thread
      operationId: getChatThread
      x-astrbot-scope: chat
      parameters:
        - name: thread_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Chat]
      summary: Delete a webchat side thread
      operationId: deleteChatThread
      x-astrbot-scope: chat
      parameters:
        - name: thread_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/threads/{thread_id}/messages:
    post:
      tags: [Chat]
      summary: Send a message inside a webchat side thread
      operationId: sendChatThreadMessage
      x-astrbot-scope: chat
      parameters:
        - name: thread_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatThreadMessageRequest"
      responses:
        "200":
          description: Server-sent chat stream or an error envelope
...
  /api/v1/sessions:
    get:
      tags: [Sessions]
      summary: List active UMOs and session status
      operationId: listSessions
      x-astrbot-scope: data
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: search
          in: query
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
        - name: message_type
          in: query
          schema:
            type: string
            enum: [all, group, private]
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
...
    ChatRequest:
      type: object
      required: [message]
      properties:
        username:
          type: string
          description: Caller-declared WebChat sender/session owner. Configured AstrBot administrator IDs require the chat:admin API key sub-scope.
        session_id:
          type: string
        conversation_id:
          type: string
          description: Deprecated alias for session_id.
        message:
          oneOf:
            - type: string
            - type: array
              items:
                $ref: "#/components/schemas/MessagePart"
        config_id:
          type: string
        config_name:
          type: string
        selected_provider:
          type: string
        selected_model:
          type: string
        enable_streaming:
          type: boolean
          default: true
          deprecated: true
          description: Deprecated compatibility field. It is used only when flags.enable_streaming is absent; otherwise flags.enable_streaming takes priority.
        flags:
          $ref: "#/components/schemas/ChatFlags"
        _skip_user_history:
          type: boolean
          description: Internal WebUI flag for edit/regenerate flows.
        _llm_checkpoint_id:
          type: string
          description: Internal WebUI checkpoint override.
        _platform_history_id:
          type: string
          description: Internal WebUI platform history override.
        _thread_selected_text:
          type: string
          description: Internal WebUI side-thread context.
      additionalProperties: false
    ChatSessionBatchDeleteRequest:
      type: object
      required: [session_ids]
      properties:
        session_ids:
          type: array
          items:
            type: string
      additionalProperties: false
    ChatSessionPatchRequest:
      type: object
      properties:
        display_name:
          type: string
      additionalProperties: false
    ChatOpenFileRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
      additionalProperties: false
    ChatMessagePatchRequest:
      type: object
      required: [content]
      properties:
        content:
          type: object
          additionalProperties: true
      additionalProperties: false
    ChatMessageRegenerateRequest:
      type: object
      properties:
        selected_provider:
          type: string
        selected_model:
          type: string
        enable_streaming:
          type: boolean
          default: true
          deprecated: true
          description: Deprecated compatibility field. It is used only when flags.enable_streaming is absent; otherwise flags.enable_streaming takes priority.
        flags:
          $ref: "#/components/schemas/ChatFlags"
      additionalProperties: false
    ChatThreadCreateRequest:
      type: object
      required: [session_id, parent_message_id, selected_text]
      properties:
        session_id:
          type: string
        parent_message_id:
          oneOf:
            - type: string
            - type: integer
        selected_text:
          type: string
      additionalProperties: false
    ChatThreadMessageRequest:
      type: object
      required: [message]
      properties:
        message:
          oneOf:
            - type: string
            - type: array
              items:
                $ref: "#/components/schemas/MessagePart"
        selected_provider:
          type: string
        selected_model:
          type: string
        enable_streaming:
          type: boolean
          default: true
          deprecated: true
          description: Deprecated compatibility field. It is used only when flags.enable_streaming is absent; otherwise flags.enable_streaming takes priority.
        flags:
          $ref: "#/components/schemas/ChatFlags"
      additionalProperties: false
...